requirements.txt
awesome-slugify==1.6.5
Django==1.11.4
django-appconf==1.0.2
django-ckeditor==5.3.0
django-contact-form==1.4.2
django-geojson==2.10.0
django-imagekit==4.0.1
django-js-asset==0.1.1
django-leaflet==0.22.0
django-maintenancemode==0.11.2
django-registration-redux==1.8
IPy==0.83
jsonfield==2.0.2
model-mommy==1.4.0
olefile==0.44
pilkit==2.0
Pillow==4.2.1
pkg-resources==0.0.0
python-magic==0.4.13
pytz==2017.2
regex==2017.7.28
six==1.11.0
Unidecode==0.4.21
